+++ src/App.css
@@ -7,7 +7,9 @@
     transition: opacity 0.5s;
 }
 
-
+.App {
+    text-align: center;
+}
 
 .footer {
     bottom: 0px;
