+++ src/routes/RequestedFriends.js
@@ -3,11 +3,11 @@ import MyPageLeftSideBar from '../components/MyPageLeftSideBar';
 import './css/MyPage.css';
 
 function RequestedFriends() {
-	return (
-		<div className="myPageContainer">
-			<MyPageLeftSideBar />
-			<div className="rightContent">ddddddf</div>
-		</div>
-	);
+  return (
+    <div className="myPageContainer">
+      <MyPageLeftSideBar />
+      <div className="rightContent">ddddddf</div>
+    </div>
+  );
 }
 export default RequestedFriends;
