+++ src/routes/css/Direct.css
@@ -1,0 +1,59 @@
+.direct-body {
+    color: #666;
+    font: 14px/24px "Open Sans", "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", Sans-Serif;
+}
+.direct-table {
+    border-collapse: separate;
+    border-spacing: 0;
+    width : 100%;
+}
+.direct-table td {
+    padding: 6px 15px;
+}
+.direct-table td {
+    padding: 6px 15px;
+    text-align: center;
+}
+
+/* .direct-table\tr:first-child\td:first-child의 속성변경. 이하 동일 */
+.direct-table tr:first-child td:first-child {
+    border-top-left-radius: 6px;
+}
+.direct-table tr:first-child td:last-child {
+    border-top-right-radius: 6px;
+}
+.direct-table td {
+    border-top : 1px solid #c6c9cc;
+    border-right: 1px solid #c6c9cc;
+    border-bottom: 1px solid #c6c9cc;
+}
+.direct-table td:first-child {
+    border-left: 1px solid #c6c9cc;
+}
+.direct-table tr:nth-child(even) td {
+    background: #eaeaed;
+}
+.direct-table tr:last-child td:first-child {
+    border-bottom-left-radius: 6px;
+}
+.direct-table tr:last-child td:last-child {
+    border-bottom-right-radius: 6px;
+}
+
+
+.cancel-button {
+    background-color: #9494ae; /* 배경색 (초록색) */
+    color: white; /* 글자색 */
+    font-size: 25px; /* 글자 크기 */
+    font-weight: bold; /* 글자 굵기 */
+    padding: 12px 20px; /* 상하, 좌우 여백 */
+    border: none; /* 테두리 없음 */
+    border-radius: 5px; /* 모서리 둥글게 */
+    cursor: pointer; /* 마우스 포인터 변경 */
+    transition: background-color 0.3s, transform 0.2s; /* 배경색과 크기 변환 애니메이션 */
+}
+
+.cancel-button:hover {
+    background-color: #3c3c42; /* 호버 시 배경색 (어두운 초록색) */
+    transform: scale(1.05); /* 호버 시 크기 확대 */
+}
