+++ src/routes/css/Home.css
@@ -1,152 +1,143 @@
 /* 전체 페이지 정렬 */
 .App {
-    text-align: center;
+  text-align: center;
 }
 
 @import url(https://fonts.googleapis.com/css?family=Raleway:400,500,800);
 
 /* 신발 카드 스타일 */
 figure.snip1200 {
-    font-family: "Raleway", Arial, sans-serif;
-    position: relative;
-    overflow: hidden;
-    margin: 10px;
-    width: 100%;
-    max-width: 330px;
-    height: 400px;
-    background: #ffffff;
-    /* 배경을 흰색으로 설정 */
-    color: #333;
-    /* 텍스트를 어두운 색으로 변경 */
-    text-align: center;
-    box-shadow: 0 0 5px rgba(0, 0, 0, 0.15);
-    font-size: 16px;
-    transition: transform 0.3s ease-in-out;
+  font-family: "Raleway", Arial, sans-serif;
+  position: relative;
+  overflow: hidden;
+  margin: 10px;
+  width: 100%;
+  max-width: 330px;
+  height: 400px;
+  background: #ffffff; /* 배경을 흰색으로 설정 */
+  color: #333; /* 텍스트를 어두운 색으로 변경 */
+  text-align: center;
+  box-shadow: 0 0 5px rgba(0, 0, 0, 0.15);
+  font-size: 16px;
+  transition: transform 0.3s ease-in-out;
 }
 
 figure.snip1200:hover {
-    transform: scale(1.05);
+  transform: scale(1.05);
 }
 
 figure.snip1200 * {
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-    -webkit-transition: all 0.45s ease-in-out;
-    transition: all 0.45s ease-in-out;
+  -webkit-box-sizing: border-box;
+  box-sizing: border-box;
+  -webkit-transition: all 0.45s ease-in-out;
+  transition: all 0.45s ease-in-out;
 }
 
 figure.snip1200 img {
-    max-width: 100%;
-    height: 240px;
-    object-fit: cover;
-    position: relative;
-    opacity: 0.9;
-    transition: opacity 0.3s, transform 0.3s;
+  max-width: 100%;
+  height: 240px;
+  object-fit: cover;
+  position: relative;
+  opacity: 0.9;
+  transition: opacity 0.3s, transform 0.3s;
 }
 
 figure.snip1200 figcaption {
-    position: absolute;
-    top: 60%;
-    left: 10%;
-    right: 10%;
-    bottom: 0;
-    padding-top: 10px;
-    background: rgba(255, 255, 255, 0.85);
-    /* 반투명 흰색 배경 */
-    border-radius: 10px 10px 0 0;
-    transition: top 0.3s ease-in-out, bottom 0.3s ease-in-out;
+  position: absolute;
+  top: 60%;
+  left: 10%;
+  right: 10%;
+  bottom: 0;
+  padding-top: 10px;
+  background: rgba(255, 255, 255, 0.85); /* 반투명 흰색 배경 */
+  border-radius: 10px 10px 0 0;
+  transition: top 0.3s ease-in-out, bottom 0.3s ease-in-out;
 }
 
 figure.snip1200 .heading {
-    font-size: 1.2rem;
-    font-weight: 700;
-    text-transform: uppercase;
-    margin-bottom: 10px;
+  font-size: 1.2rem;
+  font-weight: 700;
+  text-transform: uppercase;
+  margin-bottom: 10px;
 }
 
 figure.snip1200 h2 {
-    text-align: center;
-    font-weight: 400;
-    font-size: 1.6rem;
-    margin: 0;
-    color: #333;
-    /* 헤딩 색상을 어두운 색으로 변경 */
+  text-align: center;
+  font-weight: 400;
+  font-size: 1.6rem;
+  margin: 0;
+  color: #333; /* 헤딩 색상을 어두운 색으로 변경 */
 }
 
 figure.snip1200 h2 span {
-    font-weight: 800;
+  font-weight: 800;
 }
 
 figure.snip1200 p {
-    margin: 10px 0 0;
-    font-size: 0.9rem;
-    line-height: 1.4em;
-    opacity: 0;
-    transform: translateY(20%);
-    transition: opacity 0.35s ease-in-out, transform 0.35s ease-in-out;
-    color: #333;
-    /* 설명 텍스트를 어두운 색으로 변경 */
+  margin: 10px 0 0;
+  font-size: 0.9rem;
+  line-height: 1.4em;
+  opacity: 0;
+  transform: translateY(20%);
+  transition: opacity 0.35s ease-in-out, transform 0.35s ease-in-out;
+  color: #333; /* 설명 텍스트를 어두운 색으로 변경 */
 }
 
 figure.snip1200:hover p {
-    opacity: 1;
-    transform: translateY(0);
+  opacity: 1;
+  transform: translateY(0);
 }
 
 figure.snip1200:hover img {
-    opacity: 0.25;
-    transform: scale(1.1);
+  opacity: 0.25;
+  transform: scale(1.1);
 }
 
 figure.snip1200 a {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    z-index: 1;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  z-index: 1;
 }
 
 /* 가격 스타일링 */
 figure.snip1200 .price {
-    font-size: 1rem;
-    color: #ff4d4f;
-    /* 가격을 강조하기 위해 빨간색 유지 */
-    margin-top: 5px;
-    font-weight: bold;
+  font-size: 1rem;
+  color: #ff4d4f; /* 가격을 강조하기 위해 빨간색 유지 */
+  margin-top: 5px;
+  font-weight: bold;
 }
 
 /* 그리드 레이아웃 */
 .shoe-container {
-    max-width: 1200px;
-    margin: auto;
-    padding: 20px;
+  max-width: 1200px;
+  margin: auto;
+  padding: 20px;
 }
 
 .shoe-row {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
 }
 
 .shoe-card {
-    margin-bottom: 25px;
-    flex-basis: calc(30% - 20px);
-    /* 박스 크기를 살짝 넓힘 */
-    display: flex;
-    justify-content: center;
+  margin-bottom: 25px;
+  flex-basis: calc(30% - 20px); /* 박스 크기를 살짝 넓힘 */
+  display: flex;
+  justify-content: center;
 }
 
 @media (max-width: 992px) {
-    .shoe-card {
-        flex-basis: calc(45% - 20px);
-        /* 중간 화면 크기에서 넓힘 */
-    }
+  .shoe-card {
+    flex-basis: calc(45% - 20px); /* 중간 화면 크기에서 넓힘 */
+  }
 }
 
 @media (max-width: 768px) {
-    .shoe-card {
-        flex-basis: 100%;
-        /* 작은 화면에서는 전체 너비로 */
-    }
+  .shoe-card {
+    flex-basis: 100%; /* 작은 화면에서는 전체 너비로 */
+  }
 }
