+++ src/routes/css/MyPage.css
@@ -20,6 +20,8 @@
     margin-bottom: 15px;
 }
 
+
+
 .myPageContainer .rightContent {
     flex: 6;
     /* 오른쪽 콘텐츠가 남은 공간을 모두 차지 */
@@ -63,21 +65,21 @@
     align-items: center;
 }
 
-/* 오른쪽 영역의 모든 버튼 css */
 .myPageContainer .rightContent .btn {
     border: 1px solid #ccc;
 }
 
+
 .button-container {
     display: flex;
-    gap: 15px;
+    gap : 15px;
     margin-top: 30px;
-    margin-left: auto;
-
-    cursor: pointer;
-}
-
-
+    margin-left : auto;
+    
+    cursor:pointer;
+  }
+  
+  
 .modal-container {
     width: 100%;
     height: 100%;
@@ -89,7 +91,7 @@
     align-items: center;
     background: #00000080;
 }
-
+  
 .modal-content {
     background-color: #ffffff;
     width: 650px;
@@ -99,44 +101,29 @@
 
 .input-container {
     position: relative;
-    margin-bottom: 20px;
-    /* 원하는 간격 조절 */
-    width: 80%;
-    padding: 15px;
+    margin-bottom: 20px; /* 원하는 간격 조절 */
+    width:80%;
+    padding : 15px;
 }
 
 .input-underline {
     border: none;
-    border-bottom: 1px solid black;
-    /* 밑줄의 색상과 두께를 조절 */
-    outline: none;
-    /* 포커스 시 기본 테두리 제거 */
-    padding: 10px;
-    /* 원하는 패딩 조절 */
-    width: 100%;
-    /* 너비를 원하는 대로 조절 */
+    border-bottom: 1px solid black; /* 밑줄의 색상과 두께를 조절 */
+    outline: none; /* 포커스 시 기본 테두리 제거 */
+    padding: 10px; /* 원하는 패딩 조절 */
+    width: 100%; /* 너비를 원하는 대로 조절 */
 }
-
 .char-count {
     position: absolute;
-    bottom: -10px;
-    /* input의 밑줄 아래 위치 조절 */
-    right: 0;
-    /* 오른쪽 끝에 위치 */
-    font-size: 12px;
-    /* 글자 크기 조절 */
-    color: gray;
-    /* 글자 색상 조절 */
+    bottom: -10px; /* input의 밑줄 아래 위치 조절 */
+    right: 0; /* 오른쪽 끝에 위치 */
+    font-size: 12px; /* 글자 크기 조절 */
+    color: gray; /* 글자 색상 조절 */
 }
-
 .error-message {
     position: absolute;
-    bottom: -10px;
-    /* input의 밑줄 아래 위치 조절 */
-    left: 20px;
-    /* 왼쪽 끝에 위치 */
-    font-size: 14px;
-    /* 글자 크기 조절 */
-    color: red;
-    /* 글자 색상 조절 */
+    bottom: -10px; /* input의 밑줄 아래 위치 조절 */
+    left: 20px; /* 왼쪽 끝에 위치 */
+    font-size: 14px; /* 글자 크기 조절 */
+    color: red; /* 글자 색상 조절 */
 }
