+++ src/routes/css/RecentlyViewed.css
@@ -1,122 +1,121 @@
 /* 페이지 전체 스타일 */
 body {
-    font-family: "Arial", sans-serif;
+  font-family: "Arial", sans-serif;
 }
 
 h2 {
-    color: #333;
-    font-size: 2.2rem;
-    font-weight: bold;
-    margin-bottom: 20px;
+  color: #333;
+  font-size: 2.2rem;
+  font-weight: bold;
+  margin-bottom: 20px;
 }
 
 /* 테이블 스타일 */
 .table {
-    width: 80%;
-    margin: 0 auto;
-    background-color: #fff;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  width: 80%;
+  margin: 0 auto;
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
 .table-bordered {
-    border: 1px solid #dee2e6;
+  border: 1px solid #dee2e6;
 }
 
 .table-hover tbody tr:hover {
-    background-color: #f9f9f9;
+  background-color: #f9f9f9;
 }
 
 /* 상품 이미지가 박스에 꽉 차도록 스타일 수정 */
 .product-image {
-    width: 50%;
-    height: 220px;
-    /* 고정된 높이 설정으로 이미지가 박스에 맞게 보이도록 */
-    border-radius: 8px;
-    object-fit: cover;
-    transition: transform 0.2s ease-in-out;
-    margin: 0 auto;
-    display: block;
+  width: 50%;
+  height: 220px; /* 고정된 높이 설정으로 이미지가 박스에 맞게 보이도록 */
+  border-radius: 8px;
+  object-fit: cover;
+  transition: transform 0.2s ease-in-out;
+  margin: 0 auto;
+  display: block;
 }
 
 .product-image:hover {
-    transform: scale(1.05);
+  transform: scale(1.05);
 }
 
 /* 상품 정보 스타일 */
 .product-title {
-    font-size: 1.6rem;
-    font-weight: bold;
-    margin-bottom: 8px;
-    color: #333;
-    text-align: center;
+  font-size: 1.6rem;
+  font-weight: bold;
+  margin-bottom: 8px;
+  color: #333;
+  text-align: center;
 }
 
 .product-content {
-    font-size: 1.1rem;
-    color: #666;
-    margin-bottom: 8px;
-    text-align: center;
+  font-size: 1.1rem;
+  color: #666;
+  margin-bottom: 8px;
+  text-align: center;
 }
 
 .product-price {
-    font-size: 1.4rem;
-    font-weight: bold;
-    color: #e74c3c;
-    text-align: center;
+  font-size: 1.4rem;
+  font-weight: bold;
+  color: #e74c3c;
+  text-align: center;
 }
 
 .product-count {
-    font-size: 1rem;
-    color: #888;
-    margin-bottom: 15px;
-    text-align: center;
+  font-size: 1rem;
+  color: #888;
+  margin-bottom: 15px;
+  text-align: center;
 }
 
 /* 버튼 스타일 */
 .product-buttons {
-    display: flex;
-    justify-content: center;
-    gap: 10px;
-    margin-top: 10px;
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+  margin-top: 10px;
 }
 
 button {
-    font-size: 1rem;
-    padding: 8px 16px;
-    border-radius: 4px;
-    transition: background-color 0.3s ease;
+  font-size: 1rem;
+  padding: 8px 16px;
+  border-radius: 4px;
+  transition: background-color 0.3s ease;
 }
 
 button:hover {
-    opacity: 0.9;
+  opacity: 0.9;
 }
 
 button:focus {
-    outline: none;
+  outline: none;
 }
 
 button.variant-primary {
-    background-color: #007bff;
-    border-color: #007bff;
-    color: #fff;
+  background-color: #007bff;
+  border-color: #007bff;
+  color: #fff;
 }
 
 button.variant-success {
-    background-color: #28a745;
-    border-color: #28a745;
-    color: #fff;
+  background-color: #28a745;
+  border-color: #28a745;
+  color: #fff;
 }
 
 button.variant-danger {
-    background-color: #dc3545;
-    border-color: #dc3545;
-    color: #fff;
+  background-color: #dc3545;
+  border-color: #dc3545;
+  color: #fff;
 }
 
 /* 중앙 정렬 */
 p {
-    font-size: 1.2rem;
-    text-align: center;
-    color: #555;
+  font-size: 1.2rem;
+  text-align: center;
+  color: #555;
 }
